+++ WTraining/app/serializers/user.js
@@ -6,7 +6,10 @@ exports.signInSerializer = user => ({
 });
 
 exports.userSerializer = user => ({
-  ...this.signInSerializer,
+  id: user.id,
+  name: user.name,
+  last_name: user.lastName,
+  email: user.email,
   updated_at: user.updatedAt,
   created_at: user.createdAt
 });
