+++ WTraining/app/services/user.js
@@ -1,7 +1,7 @@
 'use strict';
 const logger = require('../logger');
 const { User } = require('../models');
-const { databaseError, conflictError, unauthorizedError } = require('../errors');
+const { databaseError, conflictError } = require('../errors');
 const { MSG_EMAIL_NOT_UNIQUE } = require('../constants/constants');
 
 exports.createUser = user =>
@@ -13,5 +13,5 @@ exports.createUser = user =>
 exports.findOneUser = email =>
   User.findOne({ where: { email } }).catch(e => {
     logger.error(e);
-    throw unauthorizedError(e.message);
+    throw databaseError(e.message);
   });
