+++ WTraining/test/constants/constants.js
@@ -4,3 +4,5 @@ exports.CONFLICT_ERROR = 409;
 exports.INTERNAL_ERROR = 500;
 exports.USER_ERROR = 400;
 exports.CREATED = 201;
+exports.OK = 200;
+exports.UNAUTHORIZED_ERROR = 401;
